Finished all but UpdateUser from superuser

+++ Bank/src/main/java/dev/jason/daos/AccountJDBCDAO.java
@@ -98,6 +98,7 @@ public class AccountJDBCDAO implements AccountDAO {
 
 	public boolean closeAccount(Account account) {
 
+		if (account.getBalance() != 0.0f) return false;
 		Connection conn = ConnectionUtil.getConnection();
 		String sql = "DELETE FROM account WHERE a_id=?";
 		try {
